showreview/shows/requirements.txt
psycopg2-binary
django
djangorestframework
PyJWT
pytest-django
pika